Renames rouge_reward to rouge
