+++ prompting/rewards/__init__.py
@@ -7,7 +7,7 @@ from .reward import (
 )
 from .code_diff import DiffRewardModel
 from .relevance import RelevanceRewardModel
-from .rouge_reward import RougeRewardModel
+from .rouge import RougeRewardModel
 from .float_diff import FloatDiffModel
 from .date import DateRewardModel
 from .pipeline import RewardPipeline
